Implemented more policies, and server end statistics
so pretty..

+++ src/TESTS/unitTest.c
@@ -29,7 +29,7 @@ void test(void *par)
     return;
 }
 
-int testHeuristic(void *arg)
+long testHeuristic(void *arg)
 {
     return (int)arg;
 }
